Need to create pidfile even when running with --no-fork
Previously, the scalyr-agent would only create the pidfile if running
without the --no-fork option.

When running in a docker container, the scalyr-agent must be run with
the --no-fork option.

This meant that no pidfile was being created, and so 'status' queries
were failing because the agent couldn't find the pidfile of the main
agent.

This fix addresses that issue by making sure the pidfile is created
regardless of whether --no-fork is specified.