Only process recently modified files
This commit addresses the problem where we scan a directory of files for
processing and that directory has a large number of files (many of
which may no longer be useful or relevant), and we then start to run in
to errors due to too many open files because all files in the directory
are kept opened.

Therefore when preparing a file for inactivity, we check
whether it has been modified recently and if so, close it.

The the value of 'recently' is defined by the configuration option:
'ignore_old_files_duration_in_seconds' in agent.json and defaults to
7200 (2 hours)