Event Sequences
This commit adds event sequence functionality to each event.

Every event being copied to Scalyr will now have two new fields: a
sequence id and a sequence number.

The sequence id is a globally unique id that creates a scope for an
ordering of sequence numbers.

The sequence number is a monotonically increasing number based on the
number of bytes read from a log file over its life.

Sequence numbers can only be compared within the same sequence id.

Each log file will receive its own sequence id.  The sequence id for a
log file can change periodically, typically when the sequence number
wraps 1TB, or upon certain errors.

The sequence_id and sequence_number are preserved across scalyr-agent
restarts.

The sequence number is determined by saving the current
sequence_number + position whenever the LogFileIterator is marked.
Whenever a request is made to read the sequence_number, the current
position is added to the previously marked position, ensuring that the
same bytes read from the same file should have the same sequence_number,
which should make it easy to check for duplicates

In terms of API, there are three new fields included in the event
object sent up to the server.  All three of these fields are at the
same level as the ts (timestamp field) for a given event:

si:  Sequence Id.  If this field is not present in an event, then the
event will use the same sequence id as the previous event.
sn: Sequence number.  The follow sequence number.  Not given if sd is
given below.
sd: Sequence delta.  If given, then this is the difference between the
sequence number for this event and the previous event.  Not given if sn
is given.