Let Linux system metrics monitor log only changing values
The Linux system metrics monitor produces a lot of metrics and logs them
a couple times a minute. Some of them change only slowly, others not at all.

This change adds a configuration option to allow non-changing metrics to
be logged less frequently. The default behavior is unchanged (log
everything on every monitor run) but you can now do

  implicit_metric_monitor: false,
  monitors: [
    {
      // Listen for syslog messages
      module: "scalyr_agent.builtin_monitors.linux_system_metrics",
      log_all_interval: 3600
    }
  ]

to only log the full set of metrics once an hour. For the rest of the
hour, only metrics whose values have changed will be logged. This cuts
down significantly on the size of the system metrics logs, which can
add up to a bunch of data as the number of hosts grows.