Ticket: AGENT-44
Don't add logs dynamically if a log path or glob already exist that match the
path.

The purpose of this is so that log entries specified in `agent.json` or
`agent.d/*.json` can override the parser and other attributes of dynamically
added logs.

This is a temporary solution.  The real solution is to allow people to override
log config options in docker labels, similar to how it works with k8s
annotations.