Add different docker images for syslog vs json
Added support for building different container images for our
docker support.  One uses syslog and the other relies on the
json log driver, copying the logs directly from the file system.
The json log driver is the preferred approach going forward
but need to support both for now.

As part of this, reworked how we build the container artifacts.
They are now executable scripts that contain the source tarballs
and Dockerfiles necessary to build their images.  You can execute
the scripts to build the images.