Ticket #AGENT-88
Fixes a problem where we were checking the existence of a key in a dict, but the
variable could sometimes be None.

I'm still not sure of the root cause of this, but I suspect it could be caused
if there is a container that uses a non-json logging driver, meaning it won't
have a log path, and the lack of the log path was causing the monitor not to
create the abovementioned dict.

I've fixed the stack trace and also added extra debug information to help
identify the root cause of the issue.