Ticket: #AGENT-71
Fixes unit tests by making sure to do a 'scan' of the test file before
attempting to read from it.  This makes sure that processing of the file isn't
"shortcircuited" because there are no bytes available.

This does not affect the main agent which does period scans of the files being
monitored.