Small images for the docker agent
This commit addresses several things to reduce the size of the scalyr-images
used by docker and kubernetes, and also corrects a number of problems with the
initial PR.

1. It uses multistage builds so that intermediate files needed for
processing/building dependencies and source files are not included in the final
image (this reduces the image by ~10MB over the original PR).  It also means
that each step of the intermediate build can be performed with an individual
`RUN` command, which improves caching during the build process.

   Note: multi-stage builds require docker version >= 17.05 for building the
   images (this has been released for over 2 years now).

2. Syslog ports are now only exposed in the docker-syslog agent.  It is incorrect to
expose them in the docker-json or the k8s agent.  To facilitate this, a new
Dockerfile was added `Dockerfile.syslog` and `build_package.py` was updated
accordingly.  I also updated the README.md so that it no longer manually exposes
the syslog port (because these are exposed automatically).

3. The build from source Dockerfile had a number of errors in it.  Firstly, the
PR removed the `git` package and so none of the git commands workd.  Secondly,
the `k8s_preview` branch referenced in the Dockerfile no longer exists so I
updated this to use `master`.  Thirdly, the docker_tarball option of
`build_package.py` no longer exists, so I changed it to use the `k8s_builder`
instead (as it seems like this Dockerfile was originally used for making preview
builds for k8s.

I've done preliminary building and testing with these Dockerfiles and it all
seems to work, but it would be good if someone else can test them too.