Make sure cached pod information get refreshed
Previously, cached pod information could be refreshed by purging all entries
from the cache, which would trigger a new query for those values.

With the addition of the cache warmer (which limits the rate at which queries to
the API can occur) purging the cache entries means that they that are not in the
cache and therefore get removed from the list of running containers until the
(possibly rate limited) API call returns.

This was causing values to be removed from the list of running containers during
one cycle of the main loop, which caused the agent to stop monitoring those logs
- only to be re-added again in the next iteration of the loo.

To prevent this from happening entries are now only purged from the main k8s
cache if they have not been queried (either by the main code or by the cache
warmer) within the `cache_purge_secs` interval.

The cache warmer now has additional logic to determine if a WarmerEntry is
'stale' and if so, it treats it the same as if the entry was not warm, by
querying out to the API.

This has the effect that:

* entries in the k8s cache are not purged so long as either the main code or the
  cache warmer are querying to see if they exist in the cache
* entries that have not been queried at all in `cache_purge_sec` seconds are
  purged (this ensures the cache isn't full of entries from
  short-running/one-off pods)
* The cache warmer will periodically re-queue and query any pods that go 'stale'
  i.e. they haven't been refreshed by an API call within the `cache_expiry_secs`
  interval.  These stale pods will still stay in the cache so long as the rest
  of the code is still querying for them, and they'll eventually be updated once
  the rate limited query returns.