Create separate ControlledCacheWarmer instance for metrics
Due to how the marking process works, we cannot use the same
ControlledCacheWarmer instances between both the metric
fetching and the log container listing.  To fix this, I
created a separate instance for the metrics.

The longer term fix is to move to a single ContainerIterator
instance used by both metrics and logs.  However, that is more
work because we have to make that abstraction thread safe.