Fix non-deterministic already_warm test failure
Testing the `already_warm` case is tricky because the test
thread is racing against the ControlledCacheWarmer thread.

In particular, once the ControlledCacheWarmer thread notices
there is a pod to warm (because it was returned from the
`pick_next_pod` method) it (A) first checks if it is already
in the cache and (B) requests it from the API.  For the
`already_warm` case, the entry must in the cache before (A).
Also, if the thread gets to (B), it will block indefinitely
until we simulate an API response.. which we don't do in this
test.  So, we need to make sure it is in the cache before (A).
The easiest way to do this is not invoke `end_marking` until
we simulate placing it in the cache, since `end_marking` is what
adds the pod to the warming list and makes it eligible to
be returned from `pick_next_pod`.

This explains the test failures where `already_warm` failed
while blocking for the ControlledCacheWarmer to be idle.