Removed batch updates
This commit removes the batch updates option for the k8s monitor.  Node and
namespace filters were only being used for the purposes of batch updates and so
they were removed as well.  Namepaces to ignore were only being used by the node
filter and so it has been removed from the cache also (but not the monitor,
where it is still used by _get_containers to ignore namespaces).