Added test feature to help investigate hung threads (#261)
This is meant to investigate the case where the test process does
not exit because it is waiting for some thread to stop that will
never stop.

To implement this, I added the ability to set a name prefix for
all StoppableThreads created.  The test framework uses this to set
a prefix including the test case, so that all threads can be tied
back to the test case that created it.

Also created a thread in the testing framework which will wait
60s for the tests to finish.  If they do not (and hence hung),
it will print out all active thread names.

Hopefully this will help give information about which test case
is creating threads that hang the tests.