Fix some bugs that resulted in tests hanging.
The primary issue was that some tests would check the current
time on a fake clock, if it wasn't great enough, wait until
the clock advanced by waiting for notifications on it.  However,
that first check was not atomic wrt to the waiting... so it was
possible for advances/notifications to be missed between the two
steps.

To fix this, we had to introduce a way to atomically check the
clock value while waiting for advances on it.