Make fixes to enable json slow encoding
I deleted one test because it relied on UTF-32 decoding to 32bit
characters, which is not a guaranteed in Python 2.7.  Unichrs
can be either 16bit or 32bit.  Since this is such a small case
and I couldn't find a way to determine if unichr where 32 or 16,
I just had to delete the test.

Medium term, we might just want to delete the slow encoding code
paths.  This JSON library should really only be used to talk to
the Scalyr server... and the Scalyr server relies on the fast
encoding code paths.  I see there are a few places in the code
where we still use this to serialize for places other than
/addEvents.  We should just move those to the standard json
libraries.