PY2->3 Changed most u'strings' to six.u( 'strings' )
Not all were changed as there seems to be some issues with \\
replacement and the implications of this need to be analysed more
carefully.

If we decide to only support python 3.3 or higher we can avoid six.u()
entirely, as 'u' was added back as a string prefix, with no purpose
other than serving to make conversion between 2->3 more simple.