Fix issue importing third_party backports
Due to pretty complicated reasons, on systems that already have
a backports module defined, we cannot override it with our own
from third_party, even if we put the third_party directory in the
front of the PYTHONPATH.

This hasn't impacted customers since only Docker relies on this
and for Docker, we control the version of Python installed in
the base image.  However, if you are running tests in your
local machine, you might hit this.

Added a hack to fix this issue.