Force handlers in root logger
To get tests to properly capture the logging output generated
by the unit tests, we must play some games with the handlers
on the root logger object.  In particular, we need to remove
all other handlers everytime we reset the logging to use
stdout.  Otherwise, other loggers created by random acts of
code, will linger around and confuse the output.