Merge Kubernetes Sidecar Mode for High Volume Logging (#331)
This adds a sidecar mode that is intended for use when high volume log throughput is required, at a higher rate than can be sustained by a single agent.

In sidecar mode, the Agent runs as an additional container, within the same Pod as the workload for which logs are being collected.

This mode is controlled with the environment variable `K8S_SIDECAR_MODE=True`. The default is disabled. In the disabled mode, there is no change from the previous functionality.

Of note, when using the agent as a sidecar, overall CPU and Memory resource usage will be increased. Additionally, the agent RBAC permissions will need to be added to the service account which is associated with the Pod, if a service account is already being used. If there is no existing custom service account for the Pod, then the default agent ServiceAccount can be used.