Fix intermediately failing test.
Make sure the max_write_burst aka bucket_size is always larger than
the len of the formatted first and second message.

If that's not the case, tests may fail (depending on the timing and
the bucket fill rate).

Keep in mind that we take the whole formatted message into account and
not just the raw string we pass to the logger method.