Store original and formatted record size on the LogRecord object and update tests to use this value.
The affected test has been updated to use this value and fail early if
the length of the formatted first and second message exceeds
max_write_burst since this would indicate invalid test and mean test may
occasionally fail (depending on the timing and bucket fill rate).

This was the case before the fix in this branch.

Now with this change, the tests will fail fast, consistently and loudly if
that occurs (e.g. if format string changes or similar change which could
affect that code / tests).