Improved parsing of JSON and CRI long log lines
Introduced concept of "extended lines" in the LogFileIterator.
These are used when we are using an internal parser (such as json)
but the underlying line is longer than the max line size.

At the cost of using extra memory, customers can now change settings
to read the entire JSON object and then break the resulting log
line extracted from the JSON object into smaller fragments less
than the max line size.

Had to introduce a buffer to hold the extended line content as well
as a secondary position object to point at which fragment in the
extended line content should be returned next.