Send actual commit date with each CodeSpeed result.
This way we work with the correct commit date from the beginning.

If date is not provided as part of the result, CodeSpeed will use
current date and only fix / populate it with the actual correct revision
date once it fetches all the revisions for a repo (aka when user views
info for a particular revision).