Update the code to call new "existing_platform()" method instead.
This method instantiates new instance of the PlatformController class on
the first invocation. On subsequent ones, it returns existing cached
instance.

This way we avoid unnecessary object churn when instantiating
Configuration object (aka on each config load / reload).

NOTE: This shouldn't introduce a cycle or a memory leak, but it's worth
keeping an eye on it.