Various small optimizations (#467)
* Speed up json object parsing in the agent.

Instead of iterating over the existing object and creating new object
with filtered and processed attributes we manipulate the existing
object.

That approach is many time faster. Actual speed gain depends on the
actual object size (aka number of keys - the more keys input JSON object
has the bigger the difference).

* Add __slots__ attribute to the LogLine class.

This allows for more efficient attribute acces and memory usage.

Example sizes with and without slot for LogLine object with a small
"line attribute": 1056 vs 888 bytes.

* Add support for orjson library which appears to be faster for objects
which are not very large.

Json library used can now be configured via agent config file. For
backward compatibility reasons it defaults to "auto" (try to use ujson
and if it's not present fall back to built in stdlib json library).