Update dockerized decorator so it correctly handles artifacts copying.
We run each test case in a separate container which means we need to
preserve log file for each test case run (aka each container).

Previous code didn't take this into account so it only copied /
preserved agent.log file from the container which belonged to the last
test run.

In addition to that, update the code to support preservation of
arbitrary number of file paths.