Disable default conf files in Debian
For some reaosn, if `/etc/init.d/scalyr-agent-2` is marked as
a conf file, that file is not created during the install
process now that it points at an another conf file.  It is
not clear why that is.  However, it was not the intention
for `/etc/init.d/scalyr-agent-2` to be a conf file, so fixing
that resolves the issue.

Of course, if there were any customers relying on the fact that
`/etc/init.d/scalyr-agent-2` happened to be a conf file, we are
going to break them.