Enable coverage collection for AgentRunner which use dev install.
We also combine the coverage for the whole process after stopping. This allows
us to easily retrieve a single .coverage data file for that test run. Keep in
mind that we will still need to combine all the coverage files for all the test
function runs together at a later stage. This should happen somewhere inside the
Circle CI job or similar.

Also keep in mind that we will likely need to do some path processing for the
final combine to work since different tests use different paths and the paths need
to be the same for combining process to work correctly (e.g. /agent_source vs
/home/circleci/scalyr-agent-2 vs /usr/share/scalyr-agent-2/py/, etc.)

Also print any stdout and stderr produced by the agent process after stopping it.
This allows us to identify issues which happen early on before forking which don't
make it into agent.log file.