Update code and introduce another speed-optimized approach for parsing RFC3339 date strings which utilizes "udatetime" 3rd party library.
Based on the benchmarking results, this approach seems to be the fastest
(udatetime -> string split -> regex -> strptime from the fastest to the slowest
approach).

This library is only available for posix platforms and utilizes C extension
which means we still need a "fast" fallback on platforms where it's not
available.

In such scenario, we fall back to second fastest approach (aka string.split
one).

For better organization, move date parsing related utility functions in a new
module and update affected code.

Also add new tests and benchmarking code.

Keep in mind that the only reason we include "private" underscore prefixed
functions for conversion is for benchmarking and auditability.