Remove monotonically increasing timestamps from the agent (#AGENT-337)
This commit removes code from the agent that ensures all logs have a
monotonically increasing timestamp.

Now, if the agent is in standalone mode, timestamps will be set to the current
time when reading bytes from disk, or if the agent is in k8s, docker or any
other mode that parses lines as json, then the agent will try to extract a
timestamp from the json and use that.

The previous behavior can be turned back on by settings the configuration
option `enforce_monotonic_timestamps` to `True`.  The default is `False`.

Previously when parsing docker/k8s lines that included a timestamp, the agent
would add the raw timestamp to an event attribute called `raw_timestamp` so
that the original timestamp could still be accessed, even if the event
timestamp was later altered.  Now that we are no longer altering timestamps,
this field is no longer necessary, however it is currently being left in for
backwards compatibility purposes.  You can prevent this attribute from being
included by setting the configuration option `include_raw_timestamp_field` to
`False` (it defaults to `True`.