Print useful config settings to agent.log
This commit will cause the agent print various useful configuration options to
`agent.log` at startup or whenever the configuration changes.

If triggered by a config change, only those options that have changed are printed.

This will be useful in determining the values of the configuration options when
diagnosing customer issues