Add namespace whitelist capability for Kubernetes (#529)
Using this capability, users can now configure the Scalyr Agent to only collect logs from pods that belong to a specific set of namespaces.  Previously, we only allowed customers to specify which namespaces to exclude.  This provides the complementary whitelist capability.

The whitelist is set via the `k8s_include_namespaces` agent configuration file option, or the `SCALYR_K8S_INCLUDE_NAMESPACES` environment variable.  The value should be a comma separated list of the namespaces to include.  A single value of `*` means all namespaces (which is the default value).

The namespaces to exclude are applied after the whitelist.  So, if `k8s_include_namespaces` contains a namespace like `foo`, which is also included in `k8s_ingore_namespaces`, then `foo` will  not be included.