AGENT-388 Add extra_config_path as an option
Under kubernetes, it is not convenient to specify custom configuration snippets
in the agent.d directory without overriding core configuration snippets.

This PR allows users to keep the default configuration snippets we provide, but
also provide their own configuration snippets in an additional configuration
directory.

This additional configuration directory is specified via an environment
variable or a command line parameter.

If present, the agent will process all .json files in this directory and add
them to its current configuration.

This only applies to the main agent, and not the agent config scripts.