AGENT-270/373 - Ensure we don't ignore short-lived pods
Problem
--

When the k8s monitor is using the docker socket for querying containers,
it misses out on short-lived pods.

The reason this happens is that when we query the docker socket, we
query all containers (running and stopped), but filter them by
containers started after the previous time we queried the containers.

This allows us to pick up short-lived containers that started and
stopped since our previous query.

However, since the k8s ratelimiter changes were added, if no matching
pod is found in the k8s cache then the container is also skipped. This
will always happen the first time a pod is queried on the cache, and so
it means that short-lived pods are now ignored, because they won't show
up when we next query the list of running containers, because it will
only show containers created after our previous call.

Solution
--

I've added a field `has_been_used` to the WarmingEntry of the cache
warmer, which is set to False upon creation of the WarmingEntry, and a
mechanism for querying this field `pending_first_use`, which returns the
opposite of `has_been_used`.

Now when we query containers, if a container is "old and dead" but the
warming entry for the container is "pending first use" then we continue
processing the container until it has been marked as used.

We then only mark the entry as 'used' when we have read it from the cache.

This ensures that we don't ignore any containers of interest until they have
been queried from the cache at least once.