[WIP] Shard Logs Across Multiple Copying Managers
Problem
---

Due to limitations with the current ingest API, the Scalyr Agent can
only have a single request in flight for a given Scalyr Agent session,
and a single request can have a maximum of 6MB of data after
decompression.

This places a limit on the amount of data a single agent can upload to
the server, and a number of users are now running in to this limitation.

This branch is a draft to experiment with having multiple
CopyingManagers, each monitoring a subset of logs, and each with its own
connection and session.

This should allow us to have multiple requests in flight - each from a
different session.

Proposed Solution
---

Create a CopyingManagerFacade with an interface identical to the current
CopyingManager.  The rest of the application interacts with the facade
as normal, and the facade is then responsible for managing the
complexity of managing multiple CopyingManagers and sharding logs
between them.

This initial commit doesn't contain anything except a stub for the
CopyManagerFacade.  The main purpose of this commit is to create a draft
PR for design discussion.