Fix StdoutFilter.
The filter wouldn't filter correctly if a custom value was provided for
stdout_severity configuration option because getattr(logging, 'LEVEL
NAME') returns reference to the logger function for that log level and
not the actual number which maps to that log level name.

For that, we need to use logging.getLevelName().