Make sure we call .upper() on the stdout severity string.
This way it works correctly when using getattr(logging, level) and we
don't accidentaly retrieve reference to some other logging module
variable which doesn't represent a log name.

For example, if user would pass in info, we would assume the level is
correct, but we would actially get a reference to logging.info function
and not the corresponding log level number.