Don't set last_success time if we failed to send data to Scalyr due to error.
Previously we always set it, even if we failed to send data or establish
a connection which would make agent status report to be incorrect in
case we never successfuly established a connection.

NOTE: I don't know yet how this affects our dropping of stale events -
we want to make sure it doesn't negatively affect it in any way.