Bug/duplicate upload due to race (#607)
Fixed somewhat subtle bug that would lead to a log file being uploaded multiple times.  The bug is triggered when pipelined requests is enabled.  Essentially, if the list of log processors changes between when an `/addEvents` request is created and when its response is processed, then we could associate the wrong callback with the wrong log file.  In some cases, this results in a log file mistakenly recorded as being "done".  During the next scan, this "done" file would be noticed again, appearing to the Scalyr Agent as if it was a completely new log file.. resulting in that that log file being completely reuploaded again.

The fix is pretty straight-forward.  Use a key for mapping the callbacks to its log processor that is stable even if the list of log processors changes.