Fix a possible race condition in the Docker monitor which would cause, in some situations, for the monitored log to stop being monitored on restart.
This could happen if the actual container restart takes longer than the
container_check_interval or if the container check routine runs just
after the container has been restarted (stopped) and before being
started back up.

Because of the way how Docker container works (it scheduled logs for
removal), but doesn't remove them immeditely, it means new log config
would not get added when the container started back up, but it would
eventually be removed when remove logs scheduled for removal routine
would run which would mean we would incorrectly stop being monitored.

There are multiple possible solutions / approaches here. The wont
I took is to add new "force_add" argument to the CopyManager method and
if that method is true and log is scheduled to be removed on add call,
we will cancel the scheduled removal and continue to monitor the log.

This approach should also be the safest since we only pass
force_add=True in Docker monitor and as such, won't affect any other
code or monitors.