Update syslog monitor and add new "tcp_unlimited_buffer_size" config option.
For backward compatibility reasons, it defaults to True.

When this value is True, we will support frames / messages of arbitrary
sizes and "tcp_max_buffer_size" setting will tell how much bytes we try
to read from the undelrying socket in a single recv() call.

Also improve warning message we log if that condition is reached and add
tests for it.