Don't automatically run --fix-config-permissions step on Windows.
This could break the installation for users who run the agent using a
non-administrator account or account which is not part of the
Administrators group.

Sadly because the installer runs as an administrator, we can update
that fix permission script to simply detect under which user the
installation script is running to also grant that user read permissions
to the config.

For now, we manually document the fix and going forward, we may enable
it by default depending on the feedback from Windows users.