Add support for "compression_type: none" config option which allows users to disable compression all together for the outgoing requests.
This option should only be set if advised so by Scalyr support.

Currently one of the main bottlenecks in the agent is compression
operations and in some scenarios (large enough lines), disabling
compression all together can increase the throughput by up to 2x.

Do keep in mind that this of course results in more egress traffic
which may mean additional chargers for the user depending on their
provider, where their infra is located, etc.