Configure journald with globs on multiple fields (#741)
* Configure journald with globs on multiple fields

The current journald monitor only allows defining separate log
configurations based on the `unit` field.

This commit makes it possible to define log configurations based on
globs of multiple journald fields.  This is useful when docker has been
configured to use the journald log driver, because in this case the
`unit` will always be `docker`, and it is therefore preferable to match
configuration stanzas based on other fields.

The way this is handled in this commit is through the `journald_globs`
field of the journald log configuration stanzas.

`journald_globs` is a dict of globs, keyed by journald fields.  Only
fields defined in the journald monitor option `jourald_fields` are
available for glob matching.

In order to match the configuration stanza, all fields defined in
`journald_globs` must be present in the journald record, and all globs
must match the value of the field.

It is an error to define both `journald_globs` and `journald_unit` for a
single configuration stanza.  It is ok to use both of them individually
in separate configuration stanzas.

* Bail out early if regex is .* aka match all.

* Log a warning message in case we fail to parse journald_logs.journald_unit and fall back to journald_logs.journald_globs.

* Better handle value not being set error (previously we would try to call
.keys() in None which would throw an exception).

* Fix lint.

* Also add tox target for Python 3.9.

* Fix failing test.

* Add unit tests for __verify_required_attributes() method.

* Add changelog entry.

* Update agent to throw more user-friendly error which include file path
in case CA bundle doesn't exist.

* Add missing logs_field function argument for a more friendly and
correct error message.

* Fix python 2.7 compatibility.

* Add additional assertions.

Co-authored-by: Imron Alston <imron.alston@elenium.com>
Co-authored-by: Tomaz Muraus <tomazm@sentinelone.com>