Improve "skipping copying log lines" warning edge case handling (#728)
* Update agent main to not emit a warning if last total bytes produced
rate is 0.

This would indicate that the agent has just started up or similar and
bytes were skipped due to the staleness so we should not log this error
which doesn't relate to stalness but agent not being able to keep up.

* Add changelog entry.

* Include more decimal places in the warning messages so they don't end up
being confusing (e.g. showing 0.0 whereas actual value is 0.01).

* Lowercase all the header key names since casing is not critical.

* Sort the keys.

* Connection key is not returned anymore.

* Remove change.

* Add additional info.

* Update check.

* Fix failing test.

We were checking on a specific fixed mysql version, but we didn't pin
version of the MySQL server we install in the Dockerimage.

For a fix I went with a more lose regex check where we don't check on a
specific fixed patch version.

* Divide by float so the end number is a float and test passes under
Python 2.

* Fix the check.

* Opt into static IP functionality of Circle CI for AMI tests jobs so we
can configure security group whitelist.