Don't include orjson and ujson dependency in armv7 Docker image.
There is no pre-built armv7 wheel for those dependencies and for the
time being, we don't want to build them as part of the image build
process since it's slow (we need to install rust, etc).

THose two dependencies not being available means we simply fall back to
stdlib json implementation.