Add new module level monitors manager function for retrieving a reference to the currently running monitors manager instance (singleton).
This value is updated each time new active config is re-created.

Those methods can be used in other places (e.g. monitors) to retrieve a
reference to the currently active MonitorsManager without creating a
cycling dependency (MonitorsManager already refernces a list of all the
running monitors so we don't want to store reference to the
MonitorsManager directly on the Monitor class).