Add support to the date utils related functions for parsing date strings with timezones.
NOTE: Non udatetime based functions which don't support timezones (and
assume UTC) have been heavily optimized for performance (this function is
called for every single log line for Docker containers) and in case of
a timezone we fall back to dateutil library which will likely heavy non
trivial overhead in those situations.

In the future with frozen binaries we should just bundle and use
udatetime everywhere to simplify the code paths and ensure good
performance.